split and merge graphql schemas and resolvers
