+++ server/package.json
@@ -30,11 +30,12 @@
     "body-parser": "^1.18.2",
     "cors": "^2.8.4",
     "express": "^4.15.2",
-    "graphql": "^0.12.3",
+    "graphql": "^0.11.7",
     "graphql-server-express": "^1.3.2",
     "graphql-tools": "^2.14.1",
     "jsonwebtoken": "^8.1.0",
     "lodash": "^4.17.4",
+    "merge-graphql-schemas": "^1.4.0",
     "mongoose": "^4.10.8",
     "uuid": "^3.1.0"
   }
