--- server/src/graphql/query.graphql
@@ -1,15 +1,0 @@
-query {
-  allCats {
-    id
-    name
-    age
-  }
-}
-
-query {
-  loginUser(input: { username: "Jackie", password: "Chen" }) {
-    id
-    username
-    password
-  }
-}
