+++ server/src/graphql/query/mutation.graphql
@@ -51,34 +51,19 @@ mutation {
       email: "lbruce@gmail.com",
       password: "12345",
       phones: [
-        { countryCode: "+33", number: "0612345678", phonetype: WORK }
-        { countryCode: "+34", number: "0634567890", phonetype: HOME }
+        { countryCode: "+33",
+          number: "0612345678", 
+          phonetype: WORK 
+        }
+        { countryCode: "+34", 
+          number: "0634567890", 
+          phonetype: HOME 
+        }
       ]
     }
   ) {
     id
     username
     email
-  }
-}
-
-mutation {
-  login(input: { email: "jackie@gmail.com", password: "123" }) {
-    id
-    email
-    username
-  }
-}
-
-mutation {
-  login(input: { email: "jackie@gmail.com", password: "123" }) {
-    id
-    email
-    username
-    phones {
-      countryCode
-      number
-      phonetype
-    }
   }
 }
