+++ server/src/graphql/resolvers.js
@@ -1,0 +1,6 @@
+import path from 'path';
+import { fileLoader, mergeResolvers } from 'merge-graphql-schemas';
+
+const resolversArray = fileLoader(path.join(__dirname, './resolvers'));
+
+export default mergeResolvers(resolversArray);
