+++ server/src/graphql/resolvers/userResolver.js
@@ -1,6 +1,4 @@
 import bcrypt from 'bcrypt';
-// import jwt from 'jsonwebtoken';
-// import _ from 'lodash';
 
 const resolvers = {
   Query: {
@@ -38,7 +36,6 @@ const resolvers = {
         }).save();
         return phone.id;
       }));
-      console.log(phoneIds);
 
       // save user
       password = await bcrypt.hash(password, 12);
@@ -50,31 +47,6 @@ const resolvers = {
       }).save();
       return registered;
     }
-    // login: async (root, { input }, { User, Phone }) => {
-    //   const { email, password } = input;
-    //   const user = await User.findOne({ email });
-    //   console.log(user);
-    //   if (!user) {
-    //     throw new Error('No regiestered user with that email');
-    //   }
-
-    //   const valid = await bcrypt.compare(password, user.password);
-    //   if (!valid) {
-    //     throw new Error('Incorrect password');
-    //   }
-
-    // const token = jwt.sign(
-    //   {
-    //     user: _.pick(user, ['id', 'username'])
-    //   },
-    //   SECRET,
-    //   {
-    //     expiresIn: '20s'
-    //   }
-    // );
-
-    // return user;
-    // }
   }
 };
