+++ server/src/graphql/schema.js
@@ -1,0 +1,5 @@
+import { makeExecutableSchema } from 'graphql-tools';
+import typeDefs from './typeDefs';
+import resolvers from './resolvers';
+
+export default makeExecutableSchema({ typeDefs, resolvers });
