+++ server/src/graphql/typeDefs.js
@@ -1,0 +1,6 @@
+import path from 'path';
+import { fileLoader, mergeTypes } from 'merge-graphql-schemas';
+
+const typesArray = fileLoader(path.join(__dirname, './types'));
+
+export default mergeTypes(typesArray);
