+++ server/src/graphql/types/enums.js
@@ -1,0 +1,11 @@
+export default `
+  enum PhoneNumberType {
+    WORK
+    HOME
+  }
+
+  enum AddressType {
+    MAIN
+    SUBS
+  }
+`;
