+++ server/src/graphql/types/phoneTypes.js
@@ -1,0 +1,8 @@
+export default `
+  type Phone {
+    id: ID!
+    countryCode: String!
+    number: String!
+    phonetype: PhoneNumberType!
+  }
+`;
