--- server/src/resolvers.js
@@ -1,25 +1,0 @@
-const channels = [
-  {
-    id: 1,
-    name: 'soccer'
-  },
-  {
-    id: 2,
-    name: 'baseball'
-  }
-];
-let nextId = 3;
-export const resolvers = {
-  Query: {
-    channels: () => {
-      return channels;
-    }
-  },
-  Mutation: {
-    addChannel: (root, args) => {
-      const newChannel = { id: nextId++, name: args.name };
-      channels.push(newChannel);
-      return newChannel;
-    }
-  }
-};
